[1194] Add provider count cookie
We currently do a request for the whole list of providers when we hit
the organisation#show page. This is fairly inefficient, and moreover,
we'll need this in the future on every page that has breadcrumbs (almost
all of them).

When initialising a user session, query the list of providers in order
to get the total count and save it to a cookie. This can then be referenced
from the controller/view templates in order to decide whether we need
to show the breadcrumbs.

Note that this call in the session is still quite slow (for a God user),
so it slows down initial signin, but it's only done once. We can choose
to further optimise it down the line.