[1217] Fix relation between provider and course
Before we were trying to map them with `provider_code`, which isn't an
attribute on course. The API returns the `provider_code` as
`institution_code` instead.

I've fixed these and fixed all the tests that failed by doing so. Most
of them were due to explicitly looking providers up by `provider_code`.

This allows us to look up a provider and do `provider.courses`, assuming
that we included courses in the provider lookup.

Likewise, we can do `course.provider`, assuming the provider was
included in the course lookup.