[1369] Update feature specs to work with pre-populated Google forms
Because the pre-population uses the user's email, it was necessary
to change the tests to do real-ish authentication, otherwise the user
info isn't available in the views.

The specs also had to be changed from deep-linking to actually following
the path that the user would take through the service, since the app currently
doesn't support deep-linking to specific pages when unauthenticated. This needed
stubbing to be added to handle the extra journeys through the intermediate pages.