[1302] fix authorisation errors caused by race condition
Apparently Faraday is not thread-safe and the way we were setting the
authorization token on the connection was opening us up to a race condition.
This plugs that hole by using the Thread key-value context to save the user
token, and using that when we generate the connection to mcb.