[1419] Set multiple providers after authenticating
Previously, we'd run this on a callback. This caused it to get called
multiple times, and before authentication was done.

As a result, it never really got set, because it was being stored in the
session under the `auth_user`, which was `nil`. The exception thrown was
then just swallowed silently.