[1458] Make factories able to render JSONAPI
This automatically creates JSONAPI serialisers for our factories. This
is currently not done for any factory, but we'll be able to whitelist
them one by one to make this happen.

This also defines `#to_jsonapi` on all models that are built. That
method creates a JSONAPI renderer and uses the serializer to render a
real JSONAPI hash.

TODO: Make this support `includes`
TODO: Make this support `build_list`