Expose email address for copy and paste
Some users struggle to use mailto links, either they don't have an email client set up on their machine or they don't know how to get the email from the link.

Expose the email address so it's easier to copy and paste for everyone.

The address is long, so allow it to wrap before the '@'