[1610] Add provider code to course factory
Due to the way we use the factories, it's difficult to set the
provider_code on the course without causing an endless loop.

Also, setting provider_code on the course factory gets overwritten by
the provider being set to nil in the factory, and removing that breaks
things.

This is further broken by a bug in the JSONAPI client.