[1624] Display error summary at top of enrichment pages
Renders the shared errors template at the top of the about, fees,
salary, and requirements pages.

The forms are updated to each contain a hidden field, for the `page`
that they correspond to, to allow the `courses#update` action to render
the correct page when responding with validation errors.

4 new `update` routes are created, one for each page. This is because on
the regular `update` endpoint, the URL is
`/organisations/ORG/courses/CRS`. If a user takes this URL and submits
it in the same browser tab, they'll be taken to the basic details, which is
confusing.