[1671] Mark fee attributes on course as strings
This prevents json_api_client from typecasting a valid value of ''
(empty string) to `0`, which are two separate values in our system. It
also removes some weird behaviour, like how `9,000.00` is cast to `9`.

We should revisit our own casting and validations to make sure we're
doing as much work as we can so that users don't constantly bump into "X
is not a number" type errors.