Only list new or running sites in preview
Whereas on live we want to show only running sites, on preview the course might not be live – when in this state all sites will be in status "new".

Sites that are new or running but have no vacancies should show, but the vacancies column should say "No"

Increase test coverage to check for running, new and suspended sites, with and without vacancies.