[1664] Handle publish errors better
This makes the `@course.publish` method behave more similarly to
`.save`, where it automatically populates `@course.errors` with the
correct hash of errors. This is only possible now that the backend gives
us back `pointer`s on the error object returned from `publish`.

There will be follow-up work to change this from using a flash to doing a
`render` instead.