[1938] rename SitePrism's Courses to CoursesPage
Specifically, PageObjects::Page::Organisation::Courses is being renamed as we're
going to be adding a bunch of page objects for page editing/creation, and could
really benefit from putting everything related to courses in a subdir. We can't
do this without turning `Courses` into a module, which means we need to move the
existing class to `CoursesPage`.

Side note: we should probably use the suffix Pages on all page objects, and
maybe tidy up by removing some of the redundancy from the full path.