[2007] Remove single location special case
For some of our courses there is bad data – some don't have any locations. We can't easily set locations for these course.

We were assuming that for providers with 1 location, all courses would have that location. That's not the case.

Most of the affected courses weren't running in 2019.

Make it possible for providers to self-serve and fix their courses by going in and adding the missing location.

Move guidance about managing locations to the Pick your locations page and make it available to all providers.