[2117] check user is present in ApplicationController#authenticate
This is an attempt to fix an issue where users login, but are told that they
don't belong to any organisations when they actually are. This addresses the
potential issue that we're testing how current_user is being tested ... in the
view we are using "current_user.present?" whereas in authenticate we were using
simply "if current_user". This change bring them in sync (using "present?") in
hopes that that fixes the issue.