[2138] Redirect users on first sign in to accept T&Cs
This should fix the current 403s we're seeing live.

We store the user's initial transition state as part of the session, and
decide as part of the sessions create where to initially send them after
fully initializing the session.

We also need some additional plumbing to decide where to send users
after they accept terms, whether it's directly to the app, to where
they were initially going to, or to another screen.