[2405] Show no-organisations page if user has no user record
The backend returns 401 if there is no user account. The last cleanup of
the status handling code left this user journey showing the "something
went wrong" unauthorized page.

This change means that the same page is shown regardless of whether you
have no account in our database (401) or have an account but no access
to any providers (403), but the response codes to the browser are
reflective of which code the backend responded with (401/403).