[2326] Refactor - improve readability of controller's language manipulation
PR feedback.
This should make it easier to understand the reason for re-combining two
lists of subjects.

In summary:

We are modifying the list of language subjects, but leaving the other
types of subjects intact.  Because it's a replacemend of the whole
subject list we need to reconstruct the combined result.