[3054] Fix missing language selections when editing new course
Users editing modern languages selections from the new-course
confirmation page were not seeing their previous selections.

The subject parameters were being overwritten by `build_course_params`
during the "continue" (302 redirect) step of the subjects controller
resulting in loss of previously selected modern languages when arriving
on the modern languages selection page.

Modifying `build_course_params` in `subjects_controller` fixes
remembering what was selected. However, fixing that meant de-selected
languages weren't removed so `build_course_params` in
`modern_languages_controller` had to be updated too.

It is not ideal that there is now two more expensive API calls, but
without these or a significant change to the API the information as to
which subjects are languages is not available when we need it.

Co-authored-by: De Fong <de_fong2k@yahoo.com>
Co-authored-by: Faissal Bensefia <faissal@madetech.com>