[3054] Remove "going back to modern languages page" spec
PR feedback was that the comment about russian being clicked but then
not being in the assert was confusing.

It seems that this spec was testing the resultant params, but these are
supplied by the API which is entirely mocked out here so this spec
doesn't seem to entirely make sense.

The assertion that a particular route is the final state should be a
concern of the step logic not this spec.

Given that we can't figure out for sure what regression this test was
intended to prevent we are removing it.

Co-authored-by: Faissal Bensefia <faissal@madetech.com>