[3082] Style admin only features
Page sections for admin only features should have a purple left border,
indented and have a section title of "Admin section".

If a navigation link is only available for admin users then the css class
of `app-admin-only` should be added to the link. This will prefix a paddlock
to the link.

As Find Teacher Training development team are the only admin users, the
styling hasn't been researched or made a priority. This is merely to help
the team easily identify features that normal users do not have access too.