[3149] Get courses as an accredited body in one request
The previous way of getting the courses was too slow, which stopped us from
releasing. Since the new course endpoint was added to v2 in
https://github.com/DFE-Digital/teacher-training-api/pull/1288, we can now get
the courses in a single request. This significantly reduced the request time
(by approx 19 seconds for UCL).

This can be optimised further when support for sparse fieldsets is added to v2.