[3204] Update status tags design
This brings status tags inline with GOV.UK Design System
(https://design-system.service.gov.uk/components/tag/). It also means
that we can remove some of our custom css. Unfortunately we cannot remove
all of it as we discovered that on the course details page, we use the
tags on a grey background. The tagds that use grey backgrounds then end
up blending in with the background. To resolve this issue we decided to
place a 1px solid border around the tag and the border colour matches the
text colour. This custom styling of the tags should only be applied where
tags are used on a non-white background.