[3204] Remove unused css variable and class name
`$button-colour` was only previous used for publish status tags but since
thats now been removed there is no need for this variable.

`phase-tag--small`, I could not find and css for this class name. I removed
it and checked the pages where we use this code and there were no visual
differences.