[3286] Refactor allocations spec
Rename training_provider to provider, which is a provider that is not an
accredited body. Training provider will be used in reference to training
providers that have courses accredited by the accredited body. Move the api
request stub to where the training provider is build so the
click_request_pe_courses method can be reused.