[2943] Add JS testing to Rake default task
Currently we are not running javascript unit tests (Jest) as part of our
pipeline or as part of our standard development process. This commit adds
jest testing command to rake so that for the moment we will at least be
running some sort of JS testing locally before pushing to github.