[3343] add signin using magic link, implemented via the API
Add an input field to the signin intercept page to allow the user to enter an
email on which to receive a magic link, if it's recognised in the system.

Add an endpoint that can be used to sign-in a user if they have the correct
token and email address.

Both these methods use Faraday directly to sidestep the authentication mechanism
we have, which will automatically try to use "current_user" to add the
authorization header to the request. This will fail because the user isn't
authenticated. So we hand-roll the authorization request. This could be improved
by adding a service to handle the JWT token generation, I expect a PR comment to
this effect.