[3304] Add start page
This page can be optionally enabled to be displayed when a user is
unauthenticated but tries to access a login protected page.

From the start page the user can choose to click through to DfE signin.
We are also implementing magic link based signin from this page for when
Signin is unavailable.