[3405] Add a healthcheck endpoint
Co-authored-by: Alistair Laing <aliuk2012@users.noreply.github.com>
Co-authored-by: Tim Abell <tim@timwise.co.uk>

This new `/healthcheck` endpoint checks whether the dependencies of this
service are up. By implication of having returned a 200 OK status you
can also tell that this service is up.

Failure of dependencies results in a 503 service unavailable status and a json
body showing which dependencies failed.

Front-door will probably continue to check root `/` for a 200 status as
that doesn't have any dependencies currently, so no need for a ping
endpoint as well.

This code is based on
DFE-Digital/find-teacher-training#159