[3392] Add UserNotificationPreferences
Client model for the API V2 /user_notification_preferences endpoint.

Additional logic added in `explicitly_enabled` to indicate whether a
user has stored a preference previously. This combines `enabled` and
`updated_at` with a `nil` `updated_at` indicating that the user has not
previously stored preferences. This will allow the UI to reflect this.

Co-authored-by: Aga Dufrat <aga.dufrat@digital.education.gov.uk>