[3455] Add ProviderView with #show_notifications_link?
The notifications link will be displayed on the provider show page when the
provider is accredited body and when the user is associated with exactly one
accredited body. When #show_notifications_link? is false the link will be
displayed on the Organisations page, which will be implemented in the next
commits.