Make notifications admin only
A [previous
PR](https://github.com/DFE-Digital/publish-teacher-training/pull/1118)
was merged that released the notifications feature to all accredited
body users. Unfortunately there is an issue with the notifications not
being sent on course update.

This reverts that PR (Couldn't easily git revert as there were conflicts)
while we address the issue.