[3474] Remove number of places from views
As the data has a few holes in it, we decided to not display the number
of places to the user. We are aware that some training providers do not
have any recorded allocations but we will be able to identify them by
searching for repeat requests with 0 places. More work will be done
afterwards.