[3467] Add view tests and refactor
I discovered that we had feature tests setup for the recruitment_cycle
show page. All they were doing were checking the page content and they should have
really been view tests which are faster to run.

I removed "data-qa" attributes from the view as we agreed to not use these
attributes for testing. As a resulted I had to update some of the site_prism
page objects lookups to find the elements on the page.

Since `data-qa` attribute was removed, I discovered that in the views we
were using a mix of `govuk_link_to` and rails `link_to` to generate links.
I have now converted the section links from `link_to` to `govuk_link_to`.