[3668] Refactor rollover settings
This commit refactors the `rollover` setting. The name of the variable
did not accurately describe the purpose of the setting. Misha and I both
felt that this feature should be nested under the 'features' setting.
The `rollover` setting is being used to determine whether or not providers
are allowed to edit the current and next recruitment cycle which is why I
think its more appropriate to name it as `can_edit_current_and_next_cycles`.

Previously developers would use this setting like this
`Settings.rollover`

Now it can be used like this through the FeatureService
`FeatureService.require("rollover.can_edit_current_and_next_cycles")`
The above will raise an error if the feature is disabled.

or

`FeatureService.enabled?("rollover.can_edit_current_and_next_cycles")`
The above will return true/false based on the setting of the feature.