[3668] Refactor `current_cycle_open` settings
This commit refactors the `current_cycle_open` setting. The name of the
variable did not accurately describe the purpose of the setting.
Misha and I bothfelt that this feature should be nested under the 'features'
setting. The `current_cycle_open` setting is set to true for most of the
year except for a brief moment between the end of a current cycle and the
start of the next cycle. Once the next cycle begins and the `current_cycle`
setting is updated to the next cycle then this setting can revert back to `true`
I think its more appropriate to name it as `has_current_cycle_started?`.

Previously developers would use this setting like this
`Settings.current_cycle_open`

Now it can be used like this through the FeatureService
`FeatureService.require("rollover.has_current_cycle_started?")`
The above will raise an error if the feature is disabled.

or

`FeatureService.enabled?("rollover.has_current_cycle_started?")`
The above will return true/false based on the setting of the feature.