[3686] Switch to using `fields`
`fields` give us the benefit of not hard coding all the attribute names
and values. As apart of this change I had to switch the `form_with` to
use the decorated version instead of undecorated version `@provider`.

As provider.accredited body is just a sorted hash, we needed to call
`Provider.new` to convert it to a model which `fields` could use.

Co-authored-by: De Fong <de_fong2k@yahoo.com>