[3686] Fix params being passed to api
As we using `fields` method we had to update how accredited bodies params
were used. We now accept `accredited_bodies_attributes` hash from the browser
and then transform it to an array to be used as `accredited_bodies`. This
means we do not have to update the api endpoint and we can just carry on.

Co-authored-by: De Fong <de_fong2k@yahoo.com>