[3707] Fix autocomplete clearing hidden input value
This fixes a bug with the autocomplet. As soon as the users tabbed
out of the autocomplete or clicked on the submit button or anything which
meant the autocomplete lost focus, the `onConfirm` method would run.  When
it is evaluated a second time the "option" variable is null so the hidden
text field has its value set to an empty string.

This doesn't address our tech debt as to why this autocomplete is different
from all the others the application has. I did do an investigation and it
would seem as though we added this hidden field so that we didn't have to
do another provider lookup from the backend. I'd say from a data integrity and
security POV we should do a lookup from the backend and not rely on the frontend
alone. A user could edit the value or submit some xss.