Move subject_page_title from helper to decorator
subject_page_title should not have been a view helper as most of the time
Course is decorated and as course has to be passed into this method it
would make more sense to have it as a decorator.

As part of this commit I also made a small copy change for the `else`
statement to read `Pick a subject` instead of `Select a subject`. Judging
by course creation steps we should never actually end up using this because
only Primary and Secondary courses give users the option to pick subjects.