[125] Update contact feature
Add feature to allow the user to update contact details. Contacts now
each have an edit link in the view and an edit form.

Contacts are now a collection on the `Provider` model. Update existing
feature spec to support this. This spec doesn't conform to our current
testing guidelines but fixing that is out of the scope for this PR.

Co-authored-by: Avin Hurry <avin.hurry@digital.education.gov.uk>
Co-authored-by: David Kewal <david.kewal@digital.education.gov.uk>