Update copy on About this Course and requirements page
In order to keep the content on publish inline with recent updates on find this PR
tweaks the layout and copy of the about this course and requirements pages. The
requirements page has a details drop down added. The about page has a drop down
but it will render different information based on whether the course is at a uni
or a SCITT/Schools direct