[1757] Convert rollover interruptions pages to use cookies
This is a spike, so tests are missing and extra comments are added
For the purposes of review only. Do not merge.

We want to avoid adding a new state for every time a user has to 
accept and interrupt screen. It is also likely that we will break
Apart the rollover and rollover-recruitment interrupt pages this year
as the allocations period ends before rollover begins. If we wanted
to have these two pages separate and maintain states, we could get
the user stuck in an infinite loop if we had both pages enabled
At the same time: 
 accepted_rollover -> accepted_rollover_recruitment -> accepted_rollover

This PR shows how we could use cookies to replace this mechanism. We
would lose the ability to query how many users have seen these screens
and accepted them. But we could add timestamps to the user and wipe
them after every rollover if that is something we really need. Since
We know rollover and allocations happen once a year, at roughly the same
time, using cookies with a 6 month expiry takes care of wiping for us.