[1809] Use interrupt page acknowlegements model for interrupt pages
We want to avoid adding a new state for every time a user has to
accept and interrupt screen. It is also likely that we will break
Apart the rollover and rollover-recruitment interrupt pages this year
as the allocations period ends before rollover begins. If we wanted
to have these two pages separate and maintain states, we could get
the user stuck in an infinite loop if we had both pages enabled
At the same time:
 accepted_rollover -> accepted_rollover_recruitment -> accepted_rollover